+++ plugins/org.preesm.algorithm/src/org/preesm/algorithm/clustering/ParallelClusteringAlgorithm.java
@@ -167,6 +167,8 @@ public class ParallelClusteringAlgorithm implements IClusteringAlgorithm {
     // Get mergeable couple
     couples = PiSDFMergeabilty.getConnectedCouple(clusteringBuilder.getAlgorithm(),
         clusteringBuilder.getRepetitionVector());
+    // Remove couples of actors that are not in the same constraints
+    ClusteringHelper.removeConstrainedCouples(couples, clusteringBuilder.getScenario());
     return couples.isEmpty();
   }